Added some cool stuff
Updated main.py to the latest api version (v9), and probably some other stuff I don't remember. I also created a multiapi script based on the original that iterates through v6 to v9 APIs in case a code does not work on one API but does on another. This way you do not miss any codes but it is slower and strains your internet.